added template download function
I thought a simple function to download the template files would help a lot for user comfort.

Since internet access is assumed because the script is supposed to be sourced directly from github, a simple download could be enough for now.

To avoid bloating the number of lines that have to be changed when you bring out a new version, I also added the current version number into a function, so it can be called both for the download and in the FAIReator.

There is an issue left with having the download as part of the function description for FAIReator, since it is technically a different function, and I didn't solve it.

Finally, for some reason, git does not recognize that I just added lines in the top, so it at least tells me that I deleted and added everything. The only change I made inside FAIReator is replacing the hard coded version with the function.